# [Heap Trace] Perf: use hash map to speed up leaks mode 34x (IDFGH-9425)