**Previous PR:** https://github.com/espressif/esp-idf/pull/10521

This finishes the perf work on heap trace standalone. Leaks mode is now fast enough to be enabled in production, or during development!

Adds `esp_err_t heap_trace_set_hashmap(heap_trace_hashmap_entry_t *entries_buffer, size_t num_entries);`

### Motivation:
- leaks mode should be fast enough to be enabled *in production*
- users can leave this mode enabled after boot & print leaks automatically (or save a dump to disk) when a low ram threshold is hit
- if enabled _before_ boot, leaks mode is a simple tool to find ways to save heap memory. This PR makes that use case fast.

### Testing:
- tested on an esp32s3. 



# Perf:

Most importantly! using a map makes record add & remove O(1), so we don't see a performance cliff when an old allocation is freed. 34x faster!

**SPIRAM** (NUMTABLE 2500, NUMITER 2, NUMENTRY 6000, NUMREC 3000)
**total 41ms** // no trace
**total 3401ms** // standard trace
**total 109ms**  // with map

For internal ram and a more typical NUMTABLE 500, we are 1.8x faster with the hashmap. 

**internal memory:**
**total 665ms** // no trace
**total 1814ms** // standard trace
**total 1065ms** // with map

For SPIRAM, also NUMTABLE 500, I'm surprised its not as dramatic. But still ~1.4x faster.

**SPIRAM:**
**total 683ms** // no trace
**total 2118ms** // standard trace
**total 1491ms** // with map



```
    const int NUMREC = 600;
    const int NUMENTRY = 600;
    const int NUMTABLE = 500;
    const int NUMITER = 100;

    uint32_t caps = MALLOC_CAP_INTERNAL;
    heap_trace_record_t* r = heap_caps_calloc(NUMREC, sizeof(heap_trace_record_t), caps);
    heap_trace_hashmap_entry_t* e = heap_caps_calloc(NUMENTRY, sizeof(heap_trace_hashmap_entry_t), caps);
    void** table = (void**) calloc(NUMTABLE, sizeof(void*));

    for (int i = 0; i < 3; i++) {

        int64_t t0 = esp_timer_get_time();

        if (i == 1 || i == 2) {
            heap_trace_init_standalone(r, NUMREC);
            if (i == 2) {heap_trace_set_hashmap(e, NUMENTRY);}
            heap_trace_start(HEAP_TRACE_LEAKS);
        }

        for(int j = 0; j < NUMITER; j++) {
            for (int k = 0; k < NUMTABLE; k++) {table[k] = heap_caps_malloc(1000,caps);}
            for (int k = 0; k < NUMTABLE; k++) {free(table[k]);}
        }

        int64_t t1 = esp_timer_get_time();
        printf("total %llims\n", (t1 - t0) / 1000);
        if (i == 1 || i == 2) {
            heap_trace_stop();
            heap_trace_dump();
        }
    }
```