@chipweinberger 2023-02-18
This is (I think) unrelated to my PR, because it happens without the hashmap as well. 

But "totalAllocations" is not returning a sensible number when it hits 1919. Could there be a locking issue on master?


This is the exact code I'm using. ESP32-S3. 

[sdkconfig.txt](https://github.com/espressif/esp-idf/files/10772747/sdkconfig.txt)

```
    const int NUMREC = 3000;
    const int NUMENTRY = 6000;
    const int NUMTABLE = 2500;

    uint32_t caps = MALLOC_CAP_SPIRAM;
    heap_trace_record_t* r = heap_caps_calloc(NUMREC, sizeof(heap_trace_record_t), caps);
    heap_trace_hashmap_entry_t* e = heap_caps_calloc(NUMENTRY, sizeof(heap_trace_hashmap_entry_t), caps);
    void** table = (void**) calloc(NUMTABLE, sizeof(void*));

    printf("r %p\n", r);
    printf("e %p\n", e);
    printf("table %p\n", table);

    for (int a = 1917; a < 1921; a+=1) {

        for (int i = 0; i < 3; i++) {

            int64_t t0 = esp_timer_get_time();

            if (i == 1 || i == 2) {
                heap_trace_init_standalone(r, NUMREC);
                if (i == 2) {heap_trace_setup_hashmap(e, NUMENTRY);}
                heap_trace_start(HEAP_TRACE_LEAKS);
            }

            for(int j = 0; j < 1; j++) {
                for (int k = 0; k < a; k++) {table[k] = heap_caps_malloc(1000,caps);}
                for (int k = 0; k < a; k++) {free(table[k]); table[k] = 0;}
            }

            int64_t t1 = esp_timer_get_time();
            printf("total %llims\n", (t1 - t0) / 1000);
            if (i == 1 || i == 2) {
                heap_trace_stop();
                heap_trace_dump();
            }

            vTaskDelay(100);
        }
    }
```