@chipweinberger 2023-02-18
I've noticed a weird issue, when NUMTABLE is > 1919, total_allocations in the summary does not go past 1919.

This happens with and without the hashmap, and regardless of SPIRAM vs INTERNAL. It seems worse when using INTERNAL. Could there be a locking issue on master? Maybe the lock we are using does not work?