@chipweinberger 2023-02-23
This is my sdkconfig:
[sdkconfig.txt](https://github.com/espressif/esp-idf/files/10811245/sdkconfig.txt)