@SoucheSouche 2023-02-27
I see that you renamed a lot of variables. The thing is that we are using snake case for variable names is esp-idf so it would be nice if you could update them. Otherwise, I will have to commit on top of your commits to change the name again.

I leave a patch with the fixed variable names below feel free to use it :) I hope I didn't miss any.
[heap_trace.patch](https://github.com/espressif/esp-idf/files/10839640/heap_trace.patch)