@SoucheSouche 2023-02-27
`map_add`, `map_remove`, `map_find` and subsequently the static variables and function they utilize can be placed in flash by the compiler but are called from function that are specifically placed in IRAM. Everything should be placed in RAM but this means that - in the state of your PR - regardless of whether the hashmap is used or not, a substantial amount of RAM memory will be used to support the feature.

I think it would be wise to wrap the whole hashmap related logic under a config so that if the hashmap is not used, the impact of the heap trace on the RAM will remain the same.