@SoucheSouche 2023-02-28
Concerning the `1919 allocations` behavior, could it be that you simply run out of available heap? After all, you allocate NUMTABLE * sizeof(void*) bytes on the heap at the beginning of your test.

I tried the code on esp32s2 (randomly picked board), I added a `heap_caps_get_free_size(caps)` after the table calloc and got ~217000 bytes free. With a NUMTABLE set to 1000, it means that during the first iteration, I will try to allocate 1000000 bytes.

This is what I get: 
```
------- free size : 217656 ---------------

====== Heap Trace: 0 records (600 capacity) ======
====== Heap Trace Summary ======
Mode: Heap Trace Leaks
0 bytes 'leaked' in trace (0 allocations)
records: 0 (600 capacity, 215 high water mark)
hashmap: 600 capacity (21500 hits, 0 misses)
total allocations: 21500
total frees: 21500
```

Which means that 215000 bytes were successfully allocated. Given that for each allocation, there are a few overhead bytes related to the memory allocator algorithm, I would say that it adds up.