@SoucheSouche 2023-02-28
if I reduce the allocation size to 16 bytes, Then I get the expected value of total allocations.
```
------- free size : 217656 ---------------

====== Heap Trace: 0 records (600 capacity) ======
====== Heap Trace Summary ======
Mode: Heap Trace Leaks
0 bytes 'leaked' in trace (0 allocations)
records: 0 (600 capacity, 600 high water mark)
hashmap: 0 capacity (0 hits, 0 misses)
total allocations: 100000
total frees: 100000
(NB: Internal Buffer has overflowed, so trace data is incomplete.)
================================

====== Heap Trace: 0 records (600 capacity) ======
====== Heap Trace Summary ======
Mode: Heap Trace Leaks
0 bytes 'leaked' in trace (0 allocations)
records: 0 (600 capacity, 600 high water mark)
hashmap: 600 capacity (60000 hits, 40000 misses)
total allocations: 100000
total frees: 100000
(NB: Internal Buffer has overflowed, so trace data is incomplete.)
```