@chipweinberger 2023-02-28
> could it be that you simply run out of available heap?

of course! 🤦🏻‍♂️ Thanks for investigating it!