@chipweinberger 2023-03-27
no problem I hope your time away was enjoyable!

> more efficient

😀 nice!