@SoucheSouche 2023-04-03
Hi @chipweinberger!
So I modified your hash map implementation as follow:
- `hash_map` array is now an array of linked list.
- when adding an entry to the hash map, we compute the hash of the entry and add it to the corresponding list in `hash_map`

This is reducing collisions in the `hash_map` array significantly and gives better performances for less RAM used overall.
In the case when `HEAP_TRACE_ALL` is used, it gives even better results (I just put them there for comparisons, but just for your information the tests were run on RISCV architecture):

For 1000000 allocations and free, the code with no tracing executes in 5779ms, The code tracing enabled and no hash map takes  40262ms, the code with your implementation of the has map takes 20632ms and the code with the hash map using the linked list as entries executes in 8129ms.

So the first implementation of the hash map is 2 times faster than without hash map usage.
The implementation of hash map with linked list is  5 times faster.