@chipweinberger 2023-04-03
Interesting approach!

Yes in my design you the hash map has to big enough to always have spare room and remain fast.

Your approach seems like a nice simplification for the user.

Good work. I imagine that was tricky to code!