@chipweinberger 2023-04-04
Glad to hear it! 

I'm glad Espressif has taken an interest in the Console PRs. Hopefully not too much work - they're fully functional but of course "polish" takes awhile too.