@chipweinberger 2023-04-10
merged: https://github.com/espressif/esp-idf/commit/b699033ab31c9bbc4a3304dbd64dc03112497316